Migration done

+++ RestauranteCodenation.Domain/AgendaCardapio.cs
@@ -6,10 +6,10 @@ namespace RestauranteCodenation.Domain
 {
     public class AgendaCardapio
     {
-        public int IdCardapio { get; set; }
+        public long IdCardapio { get; set; }
         public Cardapio Cardapio { get; set; }
 
-        public int IdAgenda { get; set; }
+        public long IdAgenda { get; set; }
         public Agenda Agenda { get; set; }
     }
 }
