+++ RestauranteCodenation.Domain/Prato.cs
@@ -6,12 +6,12 @@ namespace RestauranteCodenation.Domain
 {
     public class Prato
     {
-        public int Id { get; set; }
+        public long Id { get; set; }
         public string Nome { get; set; }
         public List<PratosIngredientes> PratosIngredientes { get; set; }
         public string Descricao { get; set; }
         public decimal Preco { get; set; }
-        public int IdTipoPrato { get; set; }
+        public long IdTipoPrato { get; set; }
         public TipoPrato tipoPrato { get; set; }
     }
 }
