+++ RestauranteCodenation.Domain/PratosIngredientes.cs
@@ -6,10 +6,10 @@ namespace RestauranteCodenation.Domain
 {
     public class PratosIngredientes
     {
-        public int IdPrato { get; set; }
+        public long IdPrato { get; set; }
         public Prato Prato { get; set; }
 
-        public int IdIngrediente { get; set; }
+        public long IdIngrediente { get; set; }
         public Ingrediente Ingrediente { get; set; }
     }
 }
